+++ sr/bsa_cli/commands/Operations.hpp
@@ -32,6 +32,7 @@ enum class Operation
 {
   Commands,
   ExtractFile,
+  Folders,
   Help,
   List,
   Info
@@ -57,7 +58,7 @@ std::string operationToString(const Operation op);
  * \remarks There is no real reflection in C++, so we use this instead.
  * \return Returns an array containing all possible operations.
  */
-std::array<Operation, 5> allOperations();
+std::array<Operation, 6> allOperations();
 
 } // namespace