+++ sr/bsa_cli/changelog.md
@@ -1,8 +1,11 @@
 # Version history of bsa-cli
 
-## Version 0.?.? (2021-10-??)
+## Version 0.5.0 (2021-10-31)
 
 The new `folders` command to list all folders inside a BSA file is added.
+
+The new `extract-folder` command to extract all files from a folder of a BSA
+file is added.
 
 ## Version 0.4.1 (2021-10-30)